cli for execution metrics api
