+++ rd-api-client/src/main/java/org/rundeck/client/api/RundeckApi.java
@@ -1194,4 +1194,27 @@ public interface RundeckApi {
     @GET("metrics/metrics")
     Call<MetricsData> getMetricsData();
 
+    /* Execution Query Metrics */
+
+    /**
+     * Get stats on a Query of executions with all query parameters available
+     * @param project
+     * @param options
+     * @param jobIdListFilter
+     * @param xjobIdListFilter
+     * @param jobListFilter
+     * @param excludeJobListFilters
+     * @return
+     */
+    @Headers("Accept: application/json")
+    @GET("project/{project}/executions/metrics")
+    Call<Map<String, Object>> executionMetrics(
+        @Path("project") String project,
+        @QueryMap Map<String, String> options,
+        @Query("jobIdListFilter") List<String> jobIdListFilter,
+        @Query("excludeJobIdListFilter") List<String> xjobIdListFilter,
+        @Query("jobListFilter") List<String> jobListFilter,
+        @Query("excludeJobListFilter") List<String> excludeJobListFilters
+    );
+
 }
