+++ src/main/java/org/rundeck/client/tool/options/ExecutionResultOptions.java
@@ -1,0 +1,5 @@
+package org.rundeck.client.tool.options;
+
+public interface ExecutionResultOptions extends ExecutionOutputFormatOption, VerboseOption {
+
+}
